requirements.txt
streamlit
langchain
google.generativeai
python-dotenv
langchain-community
langchain_google_genai
pypdf2


